dev-python/requests removed.
New releases bundle even more libraries, which makes supporting this package
even more laborious and prone to error.  The latest releases are incompatible
with the currently packaged versions.

See mleighy if you need help porting packages with dependencies on requests to
alternatives.  Note that even though a version is now packaged in the upstream
tree it will remain masked across our sites, as it is heavily broken.